+++ docs/api/contrib/images/scatter.png
