+++ docs/api/contrib/index.rst
@@ -10,3 +10,4 @@ The ``yellowbrick.contrib`` package contains a variety of extra tools and experi
 
    boundaries
    statsmodels
+   scatter
