+++ docs/api/features/index.rst
@@ -12,7 +12,7 @@ our intent that the steering process will allow the data scientist to
 zoom and filter and explore the relationships between their instances
 and between dimensions.
 
-At the moment we have five feature analysis visualizers implemented:
+At the moment we have the following feature analysis visualizers implemented:
 
 -  :doc:`rankd`: rank single and pairs of features to detect covariance
 -  :doc:`radviz`: plot data points along axes ordered around a circle to detect separability
@@ -22,7 +22,7 @@ At the moment we have five feature analysis visualizers implemented:
 -  :doc:`manifold`: visualize high dimensional data using manifold learning
 -  :doc:`importances`: rank features by relative importance in a model
 -  :doc:`rfecv`: select a subset of features by importance
--  :doc:`scatter`: plot instances by selecting subsets of features
+-  :doc:`jointplot`: plot 2D correlation between features and target
 
 Feature analysis visualizers implement the ``Transformer`` API from
 scikit-learn, meaning they can be used as intermediate transform steps
@@ -43,7 +43,6 @@ is called which displays the image.
     from yellowbrick.features.manifold import Manifold
     from yellowbrick.features.importances import FeatureImportances
     from yellowbrick.features.rfecv import RFECV
-    from yellowbrick.features.scatter import ScatterVisualizer
 
 
 .. toctree::
@@ -56,4 +55,4 @@ is called which displays the image.
    manifold
    importances
    rfecv
-   scatter
+   jointplot
