+++ docs/index.rst
@@ -28,7 +28,7 @@ Feature Visualization
 - :doc:`api/features/manifold`: high dimensional visualization with manifold learning
 - :doc:`api/features/importances`: rank features by importance or linear coefficients for a specific model
 - :doc:`api/features/rfecv`: find the best subset of features based on importance
-- :doc:`Scatter and Joint Plots<api/features/scatter>`: direct data visualization with feature selection
+- :doc:`Joint Plots <api/features/jointplot>`: direct data visualization with feature selection
 
 Classification Visualization
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
