+++ tests/baseline_images/test_contrib/test_scatter/test_scatter_image_fail.png
