+++ yellowbrick/contrib/scatter.py
@@ -1,4 +1,4 @@
-# yellowbrick.features.scatter
+# yellowbrick.contrib.scatter
 # Implements a 2d scatter plot for feature analysis.
 #
 # Author:   Nathan Danielsen <nathan.danielsen@gmail.com>
@@ -16,7 +16,6 @@ Implements a 2D scatter plot for feature analysis.
 ##########################################################################
 
 import itertools
-from sklearn.utils.deprecation import deprecated
 import numpy as np
 
 from yellowbrick.features.base import DataVisualizer
@@ -30,7 +29,6 @@ from yellowbrick.style.colors import resolve_colors
 # Quick Methods
 ##########################################################################
 
-@deprecated("Will be moved to yellowbrick.contrib in v0.8")
 def scatterviz(X,
                y=None,
                ax=None,
@@ -93,7 +91,7 @@ def scatterviz(X,
 ##########################################################################
 # Static ScatterVisualizer Visualizer
 ##########################################################################
-@deprecated("Will be moved to yellowbrick.contrib in v0.8")
+
 class ScatterVisualizer(DataVisualizer):
     """
     ScatterVisualizer is a bivariate feature data visualization algorithm that
