+++ yellowbrick/features/__init__.py
@@ -21,7 +21,6 @@ Visualizers for feature analysis and diagnostics.
 from .pcoords import ParallelCoordinates, parallel_coordinates
 from .radviz import RadialVisualizer, RadViz, radviz
 from .rankd import Rank1D, rank1d, Rank2D, rank2d
-from .scatter import ScatterViz, ScatterVisualizer, scatterviz
 from .jointplot import  JointPlotVisualizer
 from .pca import PCADecomposition, pca_decomposition
 from .importances import FeatureImportances, feature_importances
